removing unused prototype code
