+++ ts/packages/memory/website/src/chunkingUtils.ts
@@ -95,43 +95,3 @@ function joinTitleUrlAndContent(
     result += pageContent;
     return result;
 }
-
-/**
- * Analyze chunk quality for testing and optimization
- */
-export interface ChunkQualityMetrics {
-    averageChunkSize: number;
-    chunkSizeVariance: number;
-    sentenceBoundaryRespected: number; // percentage
-    paragraphBoundaryRespected: number; // percentage
-}
-
-export function analyzeChunkQuality(chunks: string[]): ChunkQualityMetrics {
-    const sizes = chunks.map((chunk) => chunk.length);
-    const averageSize = sizes.reduce((a, b) => a + b, 0) / sizes.length;
-
-    const variance =
-        sizes.reduce((acc, size) => acc + Math.pow(size - averageSize, 2), 0) /
-        sizes.length;
-
-    // Simple heuristics for boundary respect (can be enhanced)
-    const sentenceBoundaryRespected =
-        (chunks.filter((chunk) => chunk.trim().match(/[.!?]$/)).length /
-            chunks.length) *
-        100;
-
-    const paragraphBoundaryRespected =
-        (chunks.filter(
-            (chunk) =>
-                chunk.trim().startsWith("\n") || chunk.trim().endsWith("\n\n"),
-        ).length /
-            chunks.length) *
-        100;
-
-    return {
-        averageChunkSize: averageSize,
-        chunkSizeVariance: variance,
-        sentenceBoundaryRespected,
-        paragraphBoundaryRespected,
-    };
-}
