+++ ts/packages/memory/website/src/extraction/types.ts
@@ -2,44 +2,91 @@
 // Licensed under the MIT License.
 
 import { conversation as kpLib } from "knowledge-processor";
-import {
-    PageContent,
-    MetaTagCollection,
-    ActionInfo,
-    StructuredDataCollection,
-} from "../contentExtractor.js";
+
+// Page content types
+export interface PageContent {
+    title: string;
+    mainContent: string;
+    headings: string[];
+    codeBlocks?: string[];
+    images?: ImageInfo[];
+    links?: LinkInfo[];
+    wordCount: number;
+    readingTime: number;
+}
+
+export interface ImageInfo {
+    src: string;
+    alt?: string;
+    width?: number;
+    height?: number;
+    isExternal?: boolean;
+}
+
+export interface LinkInfo {
+    href: string;
+    text: string;
+    isExternal: boolean;
+}
+
+export interface MetaTagCollection {
+    description?: string;
+    keywords?: string[];
+    author?: string;
+    ogTitle?: string;
+    ogDescription?: string;
+    ogType?: string;
+    twitterCard?: string;
+    custom: { [key: string]: string };
+}
+
+export interface StructuredDataCollection {
+    schemaType?: string;
+    data?: any;
+    jsonLd?: any[];
+}
+
+export interface ActionInfo {
+    type: "form" | "button" | "link";
+    action?: string;
+    method?: string;
+    text?: string;
+}
+
+export interface WebsiteContent {
+    pageContent?: PageContent;
+    metaTags?: MetaTagCollection;
+    structuredData?: StructuredDataCollection;
+    actions?: ActionInfo[];
+    extractionTime: number;
+    success: boolean;
+    error?: string;
+    detectedActions?: DetectedAction[];
+    actionSummary?: ActionSummary;
+}
+
+export interface WebsiteContentWithKnowledge extends WebsiteContent {
+    knowledge?: kpLib.KnowledgeResponse;
+    knowledgeQuality?: KnowledgeQualityMetrics;
+}
+
+export interface KnowledgeQualityMetrics {
+    entityCount: number;
+    topicCount: number;
+    actionCount: number;
+    confidence: number;
+    extractionMode: "basic" | "enhanced" | "hybrid";
+}
 
 /**
  * Extraction modes determine the level of content processing and AI usage
- *
- * @example
- * ```typescript
- * // Fast, no AI required
- * await extractor.extract(content, "basic");
- *
- * // AI-powered content analysis
- * await extractor.extract(content, "content");
- *
- * // Content + action detection
- * await extractor.extract(content, "actions");
- *
- * // Complete analysis with relationships
- * await extractor.extract(content, "full");
- * ```
+ 
  */
 export type ExtractionMode = "basic" | "summary" | "content" | "full";
 
 /**
  * Configuration for content extraction operations
  *
- * @example
- * ```typescript
- * const config: ExtractionConfig = {
- *   mode: "content",
- *   timeout: 10000,
- *   maxContentLength: 500000
- * };
- * ```
  */
 export interface ExtractionConfig {
     // Primary control - determines both content extraction and AI usage
@@ -125,16 +172,6 @@ export const EXTRACTION_MODE_CONFIGS: Record<
 /**
  * Input content for extraction operations
  *
- * @example
- * ```typescript
- * const input: ExtractionInput = {
- *   url: "https://example.com/article",
- *   title: "Article Title",
- *   htmlContent: "<html>...</html>",
- *   textContent: "Plain text content",
- *   source: "direct"
- * };
- * ```
  */
 export interface ExtractionInput {
     url: string;
